refactor: extract language server detection into dedicated module (ISM-359)
