+++ src/surfmon/db.py
@@ -20,7 +20,8 @@ from . import __version__
 if TYPE_CHECKING:
     from collections.abc import Iterator
 
-    from .monitor import LsSnapshot, MonitoringReport
+    from .language_servers import LsSnapshot
+    from .monitor import MonitoringReport
     from .pty import PtyInfo
 
 DB_DIR = Path.home() / ".surfmon"
